Move computation of designated initializers to common place
